feat(permission): PermissionEventHandler 인터페이스 추가 및 구현체 분리
- PermissionEventHandler 인터페이스 생성
- 기존 PermissionEventHandler를 PermissionEventHandlerImpl로 변경
