+++ src/main/java/com/simpleboard/board/permission/presentation/PermissionEventHandler.java
@@ -1,28 +1,11 @@
 package com.simpleboard.board.permission.presentation;
 
-import com.simpleboard.board.permission.application.command.PermissionCommandService;
 import com.simpleboard.board.permission.presentation.tmp.BoardCreatedEvent;
 import com.simpleboard.board.permission.presentation.tmp.BoardDeletedEvent;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Component;
 
-@Component
-@RequiredArgsConstructor
-public class PermissionEventHandler {
+public interface PermissionEventHandler {
 
-  private final PermissionCommandService permissionCommandService;
+  void handleBoardCreated(BoardCreatedEvent event);
 
-  @Async
-  @EventListener
-  public void handleBoardCreated(BoardCreatedEvent event) {
-    permissionCommandService.createPermissionPolicy(event.boardId(), event.userId());
-  }
-
-  @Async
-  @EventListener
-  public void handleBoardDeleted(BoardDeletedEvent event) {
-    permissionCommandService.deletePermissionPolicy(event.boardId());
-  }
+  void handleBoardDeleted(BoardDeletedEvent event);
 }
